pyproject.toml
[project]
name = "saev"
version = "0.1.0"
description = "Sparse autoencoders for vision transformers in PyTorch"
readme = "README.md"
requires-python = "==3.11.*"
dependencies = [
    "altair>=5.4.1",
    "beartype>=0.19.0",
    "datasets>=3.0.1",
    "einops>=0.8.0",
    "jaxtyping>=0.2.34",
    "marimo>=0.9.10",
    "matplotlib>=3.9.2",
    "open-clip-torch>=2.28.0",
    "pillow>=10.4.0",
    "polars>=1.12.0",
    "submitit>=1.5.2",
    "torch>=2.5.0",
    "tqdm>=4.66.5",
    "tyro>=0.8.12",
    "wandb>=0.19.9",
    "scikit-learn",
    "transformers",
    "pacmap",
    "umap-learn",
    "seaborn",
    "vllm",
    "together",
    "shapely",
    "kagglehub"
]

[project.optional-dependencies]
classification = [
    "scipy>=1.15.1",
]
training = [
    "wandb>=0.19.6",
]


[tool.setuptools]
packages = ["saev"]

[tool.ruff.lint]
ignore = ["F722"]

[tool.uv]
package = true
dev-dependencies = [
    "hypothesis>=6.119.0",
    "hypothesis-torch>=0.8.4",
    "pdoc3>=0.11.1",
    "pytest>=8.3.3",
    "pytest-xdist>=3.6.1",
    "pytest-cov>=6.0.0",
    "gradio>=3.36.1",
    "pyvips>=2.2.3",
    "huggingface-hub>=0.26.5",
]

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "serial",
]
